langchain-framework/requirements.txt
streamlit
langchain
langchain_chroma
langchain_community
python-dotenv
langchain_groq
langchain_huggingface
langchain_chroma
langchain_core
pypdf
chromadb
wikipedia
arxiv
python-dotenv
faiss-cpu
duckduckgo-search
SQLAlchemy
mysql-connector-python
validators==0.28.1
youtube_transcript_api
pytube
unstructured
numexpr
cassio
datasets
openai
tiktoken
PyPDF2
astrapy
gradio
langchain_experimental
boto3
awscli